adding textadept to list
Here is a proposal to potentially add my Textadept [module for rust](https://github.com/abaez/ta-rust)
to the list. The texteditor plugin was written a year ago and after a
small exodus, has been reinvigorated with commits. It currently has
the following features:

* Syntax highlighting out-of-the-box: updated by stable releases.
* Snippets: always expanding and easy to append your own.
* Code completion: light self written and currently being updated to
also use racer for the task.
* Linting: using rustfmt for lint.
* Go-to definition: using ctags generated for projects during a cargo
build executed directly in textadept.
* Syntax checking: very simple using rustc post compile executed in
textadept. Will be expanded to use oracle.
* API reference: both from std rust and cargo built projects on
textadept.