Add Debugging and Doc Tooltips Columns
Visual Studio Code's RustyCode plugin has lots of cool new features that
are not reflected on areweideyet, so additional columns should be added.
This also gets rid of the Syntax Checking column, as it's pretty unclear
what that even means, as linting is also Syntax Checking, which made the
table pretty inaccurate. I'm not sure if the column should be
"Documentation Tooltips" or just "Show Documentation". Showing the
Documentation is something the other editors are somewhat capable of too,
but only in the browser. They don't parse the documentation and render it
specifically, which is a pretty cool and distinguishing feature that VS
Code has, so it should be highlighted.