Update link to reflect actual URL
The url that was in place until now has nothing to do with the actual url. It might even be a malicious website as multiple suspicious looking scripts are blocked by uBlock when accessing the site. I do not know how critical the situation is but please change this as fast as possible.